Add files via upload
Project Title:
	Activity 2
	
Build Status:
	Complete
	
Video:
	https://www.loom.com/share/cc73461b832a470f81375540116165ca

Framework:
	.NET

Features:
	numericUpDown, labels, buttons
	
This form allows for the conversion of miles to feet and vice versa by entering the number (up to 100,000,000) in feet or miles in their respective boxes.
Once the number is entered, click on the corresponding button and the conversion will appear at the top of the form.  To clear the fields, simply
click the clear button.